refact: move internal functions to Internal package
