+++ conversion.go
@@ -2,7 +2,7 @@
 package conversion
 
 import (
-	"fmt"
+	"github.com/dl4ab/conversion/internal"
 
 	"encoding/binary"
 	"errors"
@@ -21,54 +21,6 @@ type Option struct {
 // IntToFloat32 converts int to float32
 func IntToFloat32(x int) (float32, error) {
 	return float32(x), nil
-}
-
-// Flattens [n][4]byte to [n * 4]byte
-func flattenBytes32(xs [][]byte) ([]byte, error) {
-	return flattenInternal(xs, 4)
-}
-
-// Flattens [n][8]byte to [n * 8]byte
-func flattenBytes64(xs [][]byte) ([]byte, error) {
-	return flattenInternal(xs, 8)
-}
-
-func flattenInternal(xs [][]byte, size int) ([]byte, error) {
-	ret := make([]byte, len(xs)*size)
-	for i, bytes := range xs {
-		if len(bytes) != size {
-			return nil, fmt.Errorf("expected [][%v]byte, but received [][%v]byte", size, len(bytes))
-		}
-		for j, b := range bytes {
-			ret[i*size+j] = b
-		}
-	}
-	return ret, nil
-}
-
-func unflattenBytes32(xs []byte) ([][]byte, error) {
-	return unflattenInternal(xs, 4)
-}
-func unflattenBytes64(xs []byte) ([][]byte, error) {
-	return unflattenInternal(xs, 8)
-}
-
-// unflattenInternal transforms []byte to [][size]byte
-// If the original data was []float32 then it can be transformed into [][4]byte.
-// Then, [][4]byte can get flatten to []byte.
-// This function "undo" the flatten by making []byte back to [][4]byte.
-func unflattenInternal(xs []byte, size int) ([][]byte, error) {
-	if len(xs)%size != 0 {
-		return nil, fmt.Errorf("expected the length of xs to be a multiple of %v but its size is %v", size, len(xs))
-	}
-
-	var ret [][]byte = nil
-
-	for i := 0; i < len(xs); i += size {
-		ret = append(ret, xs[i:i+size])
-	}
-
-	return ret, nil
 }
 
 // Float32ToBytes converts float32 to []byte with length 4
@@ -139,7 +91,7 @@ func Float32sToBytes(xs []float32, o *Option) ([]byte, error) {
 	for i, x := range xs {
 		bs[i] = Float32ToBytes(x, o)
 	}
-	return flattenBytes32(bs)
+	return internal.FlattenBytes32(bs)
 }
 
 // BytesToFloat32s converts []byte to []float32
@@ -147,7 +99,7 @@ func Float32sToBytes(xs []float32, o *Option) ([]byte, error) {
 // Step 2. [][4]byte to []float32
 func BytesToFloat32s(xs []byte, o *Option) ([]float32, error) {
 	var fs []float32
-	xxs, err := unflattenBytes32(xs)
+	xxs, err := internal.UnflattenBytes32(xs)
 	if err != nil {
 		return nil, err
 	}
@@ -171,7 +123,7 @@ func Float64sToBytes(xs []float64, o *Option) ([]byte, error) {
 	for i, x := range xs {
 		bs[i] = Float64ToBytes(x, o)
 	}
-	return flattenBytes64(bs)
+	return internal.FlattenBytes64(bs)
 }
 
 // BytesToFloat64s converts []byte to []float64
@@ -179,7 +131,7 @@ func Float64sToBytes(xs []float64, o *Option) ([]byte, error) {
 // Step 2. [][8]byte to []float64
 func BytesToFloat64s(xs []byte, o *Option) ([]float64, error) {
 	var fs []float64
-	xxs, err := unflattenBytes64(xs)
+	xxs, err := internal.UnflattenBytes64(xs)
 	if err != nil {
 		return nil, err
 	}
